requirements.txt
discord
meilisearch
asyncio
python-decouple